server/fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow-cpu==2.20.0
numpy==2.1.0
pillow==10.4.0
python-multipart==0.0.9

# Video Analysis Dependencies - Python 3.13 compatible 
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
websockets>=11.0

